AWS: provide the least privilege needed
We don't really want access to be lackluster across the system, so
don't allow lambda functions to do whatever they want to the DyanamoDB
table.

On the other hand, not everyone should be able to access the data, so
provide a user specifically created to scan the table and use that instead.